requirements.txt
Flask==1.0.2
gunicorn==20.0.4
python-dotenv==0.10.1
twitter==1.17.1
